server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
google-generativeai==0.8.3
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy<2.0
watchdog==6.0.0
ptyprocess==0.7.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
